changes to registry, pkce, and account screens for integrations
